requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cfgv==3.4.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.8
Django==4.2.1
django-cors-headers==4.3.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
filelock==3.13.1
identify==2.5.33
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
nodeenv==1.8.0
openapi-codec==1.3.2
packaging==23.2
Pillow==10.1.0
platformdirs==4.1.0
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
simplejson==3.19.2
sqlparse==0.4.4
typing_extensions==4.9.0rc1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
virtualenv==20.25.0
